Fixed a bug when parsing windowsize data
We were overflowing the 1-byte container for these values when adding them together
which broke things when the window size grew > 255

+++ server/server.go
@@ -371,9 +371,9 @@ func (self *connectionHandler) Handle() {
 								return
 							}
 
-							width := (255 * data[0]) + data[1]
-							height := (255 * data[2]) + data[3]
-							self.user.SetWindowSize(int(width), int(height))
+							width := int((255 * data[0])) + int(data[1])
+							height := int((255 * data[2])) + int(data[3])
+							self.user.SetWindowSize(width, height)
 
 						case telnet.TT:
 							self.user.SetTerminalType(string(data))
